Refactoring: remove duplicating code + added interface to keep some standarts in terms of post actions. This also make code more flexible
